kaz-image-captioning/requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.1
Pillow>=9.4.0
h5py>=3.2.1
scipy>=1.7.0

# Computer Vision
opencv-python>=4.5.0

# Translation
deep-translator>=1.11.0

# Optional (for audio features)
# playsound==1.3.0
